+++ craco.config.js
@@ -4,8 +4,8 @@ module.exports = {
   webpack: {
     alias: {
       "@Components": path.resolve(__dirname, "src/components/"),
-      "@Form": path.resolve(__dirname, "src/components/Form"),
-      "@Layouts": path.resolve(__dirname, "src/components/Layouts"),
+      "@Form": path.resolve(__dirname, "src/form"),
+      "@Layouts": path.resolve(__dirname, "src/layouts"),
       "@Services": path.resolve(__dirname, "src/services/"),
       "@Pages": path.resolve(__dirname, "src/pages/"),
       "@Utils": path.resolve(__dirname, "src/utils/"),
