+++ jsconfig.json
@@ -3,8 +3,8 @@
     "baseUrl": "./",
     "paths": {
       "@Components/*": ["src/components/*"],
-      "@Form/*": ["src/components/Form/*"],
-      "@Layouts/*": ["src/components/Layouts/*"],
+      "@Form/*": ["src/form/*"],
+      "@Layouts/*": ["src/layouts/*"],
       "@Contexts/*": ["src/contexts/*"],
       "@Pages/*": ["src/pages/*"],
       "@Services/*": ["src/services/*"],
