--- src/components/Form/index.jsx
@@ -1,15 +1,0 @@
-import DateField from "@Form/DateField";
-import FileUpload from "@Components/Form/FileUpload";
-import SelectField from "@Components/Form/SelectField";
-import TextField from "@Components/Form/TextField";
-import TextFieldWithUnit from "@Components/Form/TextFieldWithUnit";
-import TextArea from "@Components/Form/TextArea";
-
-export {
-  DateField,
-  FileUpload,
-  SelectField,
-  TextField,
-  TextFieldWithUnit,
-  TextArea,
-};
