--- src/components/Report/index.jsx
@@ -1,13 +1,0 @@
-import SerialNoFields from "@Components/Report/SerialNoFields";
-import CandidateInfoFields from "@Components/Report/CandidateInfoFields";
-import MedicalExaminationFields from "@Components/Report/MedicalExaminationFields";
-import LabInvestigationFields from "@Components/Report/LabInvestigationFields";
-import Remarks from "@Components/Report/Remarks";
-
-export {
-  SerialNoFields,
-  CandidateInfoFields,
-  MedicalExaminationFields,
-  LabInvestigationFields,
-  Remarks,
-};
