+++ src/components/global/index.jsx
@@ -1,0 +1,15 @@
+import PrivateRoute from "./PrivateRoute";
+import NavigationBar from "./NavigationBar";
+import QuickAccessButtons from "./QuickAccessButtons";
+import Heading from "./Heading";
+import TableView from "../reports-list/TableView";
+import UpdateReferenceModal from "./UpdateReferenceModal";
+
+export {
+  PrivateRoute,
+  NavigationBar,
+  QuickAccessButtons,
+  Heading,
+  TableView,
+  UpdateReferenceModal,
+};
