--- src/components/index.jsx
@@ -1,15 +1,0 @@
-import PrivateRoute from "@Components/PrivateRoute";
-import NavigationBar from "@Components/NavigationBar";
-import QuickAccessButtons from "@Components/QuickAccessButtons";
-import Heading from "@Components/Heading";
-import TableView from "@Components/TableView";
-import UpdateReferenceModal from "@Components/UpdateReferenceModal";
-
-export {
-  PrivateRoute,
-  NavigationBar,
-  QuickAccessButtons,
-  Heading,
-  TableView,
-  UpdateReferenceModal,
-};
