+++ src/components/new-report/index.jsx
@@ -1,0 +1,13 @@
+import SerialNoFields from "./SerialNoFields";
+import CandidateInfoFields from "./CandidateInfoFields";
+import MedicalExaminationFields from "./MedicalExaminationFields";
+import LabInvestigationFields from "./LabInvestigationFields";
+import Remarks from "./Remarks";
+
+export {
+  SerialNoFields,
+  CandidateInfoFields,
+  MedicalExaminationFields,
+  LabInvestigationFields,
+  Remarks,
+};
