+++ src/form/ReactiveFields/index.jsx
@@ -1,0 +1,7 @@
+import DisplayPhoto from "./DisplayPhoto";
+import ExpiryDateField from "./ExpiryDateField";
+import Pregnancy from "./Pregnancy";
+import Age from "./Age";
+import GovtId from "./GovtId";
+
+export { DisplayPhoto, ExpiryDateField, Pregnancy, Age, GovtId };
